requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
build==1.2.1
    # via pip-tools
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
exceptiongroup==1.2.0
    # via pytest
greenlet==3.0.3
    # via playwright
idna==3.6
    # via requests
importlib-metadata==7.1.0
    # via build
iniconfig==2.0.0
    # via pytest
packaging==24.0
    # via
    #   build
    #   pytest
pip-tools==7.4.1
    # via -r requirements.in
playwright==1.42.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.4.0
    # via pytest
pyee==11.0.1
    # via playwright
pyperclip==1.8.2
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.1.1
    # via
    #   -r requirements.in
    #   pytest-base-url
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.4.4
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-slugify==8.0.4
    # via pytest-playwright
requests==2.31.0
    # via pytest-base-url
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
tqdm==4.66.2
    # via -r requirements.in
typing-extensions==4.10.0
    # via pyee
urllib3==2.2.1
    # via requests
wheel==0.43.0
    # via pip-tools
zipp==3.18.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
